# openapi.yaml
openapi: 3.0.3
info:
  title: Arepos Server API
  description: REST API для управления моделями, узлами, связями и компонентами
  version: 0.1.1
  contact:
    name: API Support
servers:
  - url: http://arepos-server.arch.svc.cluster.local:8080
    description: Локальный сервер разработки
  - url: https://api.example.com
    description: Продакшн сервер
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/links/{id}:
    get:
      tags:
        - links
      summary: Получить связь по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '404':
          description: Связь не найдена
    put:
      tags:
        - links
      summary: Обновить связь между узлами
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkUpdateRequest'
      responses:
        '200':
          description: Связь обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '404':
          description: Связь, владелец, модель, узлы или тип связи не найдены
    delete:
      tags:
        - links
      summary: Удалить связь
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Связь удалена
        '404':
          description: Связь не найдена
...
  /api/v1/link-types:
    get:
      tags:
        - link-types
      summary: Список типов связей
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          description: Фильтр по владельцу
          schema:
            type: string
            format: uuid
        - name: name
          in: query
          description: Фильтр по имени (поиск по подстроке)
          schema:
            type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageLinkTypeResponse'
    post:
      tags:
        - link-types
      summary: Создать тип связи
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkTypeRequest'
      responses:
        '201':
          description: Тип связи создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkTypeResponse'
        '404':
          description: Владелец не найден
        '400':
          description: Неверный запрос
  /api/v1/link-types/{id}:
    get:
      tags:
        - link-types
      summary: Получить тип связи по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkTypeResponse'
        '404':
          description: Тип связи не найден
    put:
      tags:
        - link-types
      summary: Обновить тип связи
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkTypeUpdateRequest'
      responses:
        '200':
          description: Тип связи обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkTypeResponse'
        '404':
          description: Тип связи или владелец не найдены
    delete:
      tags:
        - link-types
      summary: Удалить тип связи
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Тип связи удален
        '404':
          description: Тип связи не найден
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access токен. Получите через /api/v1/auth/login или /api/v1/auth/register
  schemas:
...
    PageInfo:
      type: object
      description: Метаданные пагинации (Spring Data Page)
      properties:
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
        totalPages:
          type: integer
          description: Общее количество страниц
        size:
          type: integer
          description: Размер страницы
        number:
          type: integer
          description: Номер текущей страницы (с 0)
        first:
          type: boolean
          description: Первая страница
        last:
          type: boolean
          description: Последняя страница
        numberOfElements:
          type: integer
          description: Количество элементов на текущей странице
        empty:
          type: boolean
          description: Пустая страница
        sort:
          type: object
          nullable: true
          description: Параметры сортировки

    # Пользователи
...
    LinkUpdateRequest:
      type: object
      properties:
        sourceId:
          type: string
          format: uuid
          nullable: true
        targetId:
          type: string
          format: uuid
          nullable: true
        modelId:
          type: string
          format: uuid
          nullable: true
        ownerId:
          type: string
          format: uuid
          nullable: true
        linkTypeId:
          type: string
          format: uuid
          nullable: true
        attrs:
          type: string
          nullable: true
    LinkResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sourceId:
          type: string
          format: uuid
        targetId:
          type: string
          format: uuid
        modelId:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        linkTypeId:
          type: string
          format: uuid
        attrs:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
          description: Дата и время создания
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Дата и время последнего обновления
...
    LinkTypeRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        ownerId:
          type: string
          format: uuid
          description: Если не указан, берется из JWT токена
        attrs:
          type: string
          nullable: true
    LinkTypeUpdateRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        ownerId:
          type: string
          format: uuid
          nullable: true
        attrs:
          type: string
          nullable: true
    LinkTypeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ownerId:
          type: string
          format: uuid
        accessPermission:
          type: string
          enum: [ADMIN, OWNER, EDIT, VIEW]
          nullable: true
        attrs:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
          description: Дата и время создания
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Дата и время последнего обновления
    PageLinkTypeResponse:
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/LinkTypeResponse'

    # Правила связей
...